% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapwtf.R
\name{any_region}
\alias{any_region}
\title{Generate a gridded region anywhere}
\usage{
any_region(width = 50000, ...)
}
\arguments{
\item{width}{buffer with in metres}

\item{...}{proj4 parameters passed to internal function}

\item{pt}{lonlat pair}
}
\value{
RasterLayer
}
\description{
Generate a gridded region anywhere
}
\examples{
any_region()
}
