% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapwtf.R
\name{grat}
\alias{grat}
\title{plot a rough graticule over a projected raster}
\usage{
grat(x)
}
\arguments{
\item{x}{a projected raster}
}
\value{
side effect draws on existing plot
}
\description{
plot a rough graticule over a projected raster
}
\examples{
x <- any_region()
raster::plot(raster::extent(x))
grat(x)
}
