% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapwtf.R
\name{populate_topo}
\alias{populate_topo}
\title{Fill a projected raster with topography from the topo data set}
\usage{
populate_topo(x)
}
\arguments{
\item{x}{a projected raster}
}
\value{
the raster with topography values in it
}
\description{
Fill a projected raster with topography from the topo data set
}
\examples{
populate_topo(any_region())
}
