% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapwtf-package.R
\docType{data}
\name{topo}
\alias{topo}
\title{Topography data in integrized sinusoidal bins}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 5940422 rows and 2 columns.
}
\usage{
topo
}
\description{
Derived from Gebco 2008
}
\details{
attribute 'nrows' records the number of rows in the binning
}
\keyword{datasets}
