% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapwtf-package.R
\docType{package}
\name{mapwtf-package}
\alias{mapwtf}
\alias{mapwtf-package}
\title{mapwtf: Maps Anywhere}
\description{
Stores topography in densest form possible for entire earth, and provides some functions
 for generating a random plot in any projection with topography data on it.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mdsumner/mapwtf}
  \item Report bugs at \url{https://github.com/mdsumner/mapwtf/issues}
}

}
\author{
\strong{Maintainer}: Michael Sumner \email{mdsumner@gmail.com}

}
\keyword{internal}
